Load build system plugins from the classpath.
Rather than having these be passed as `--plugin-path` on the
command-line, which requires more complicated testing and a separate
code path for built-in plugins as compared to user loaded, just load
all plugins available on the classpath, and load the default ndk-build
and CMake plugins that way rather than explicitly loading them.

Users that want to load their own plugins will need to modify the Java
command line to include their own plugin JARs in the classpath.

Fixes https://github.com/google/prefab/issues/22
