+++ cli/src/main/kotlin/com/google/prefab/cli/Cli.kt
@@ -23,7 +23,6 @@ import com.github.ajalt.clikt.parameters.arguments.multiple
 import com.github.ajalt.clikt.parameters.groups.OptionGroup
 import com.github.ajalt.clikt.parameters.groups.groupChoice
 import com.github.ajalt.clikt.parameters.groups.required
-import com.github.ajalt.clikt.parameters.options.multiple
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
 import com.github.ajalt.clikt.parameters.types.choice
@@ -102,10 +101,6 @@ open class Cli :
     private val output: File by option(
         help = "Output path for generated build system integration."
     ).file(fileOkay = false).required()
-
-    private val pluginPath: List<File> by option(
-        help = "Path to build system integration plugin."
-    ).file(folderOkay = false, readable = true).multiple()
 
     private val platform: PlatformConfig by option(
         help = "Target platform. Only 'android' is currently supported."
