+++ docs/index.md
@@ -50,7 +50,6 @@ Android specific configuration options:
 Options:
   --build-system TEXT  Generate integration for the given build system.
   --output DIRECTORY   Output path for generated build system integration.
-  --plugin-path FILE   Path to build system integration plugin.
   --platform TEXT      Target platform. Only 'android' is currently supported.
   -h, --help           Show this message and exit
 ```
