+++ ndk-build-plugin/src/main/kotlin/com/google/prefab/ndkbuild/NdkBuildPluginProvider.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2019 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.prefab.ndkbuild
+
+import com.google.prefab.api.BuildSystemInterface
+import com.google.prefab.api.BuildSystemProvider
+import com.google.prefab.api.Package
+import java.io.File
+
+/**
+ * Service provider for the [NdkBuildPlugin].
+ */
+class NdkBuildPluginProvider : BuildSystemProvider {
+    override val identifier: String = "ndk-build"
+
+    override fun create(
+        outputDirectory: File,
+        packages: List<Package>
+    ): BuildSystemInterface {
+        return NdkBuildPlugin(outputDirectory, packages)
+    }
+}
