+++ ndk-build-plugin/src/main/resources/META-INF/services/com.google.prefab.api.BuildSystemProvider
@@ -1,0 +1,1 @@
+com.google.prefab.ndkbuild.NdkBuildPluginProvider
